+++ src/main/java/pokeraidbot/domain/errors/GymNotFoundException.java
@@ -5,7 +5,7 @@ import pokeraidbot.domain.config.LocaleService;
 import java.util.Locale;
 
 public class GymNotFoundException extends RuntimeException {
-    public GymNotFoundException(String name, LocaleService localeService, Locale locale) {
-        super(localeService.getMessageFor(LocaleService.GYM_NOT_FOUND, locale, name));
+    public GymNotFoundException(String name, LocaleService localeService, Locale locale, String region) {
+        super(localeService.getMessageFor(LocaleService.GYM_NOT_FOUND, locale, name, region));
     }
 }